GH-116 - create default folders.
Also better fix for GH-123 - catch EmailAccount.DoesNotExist. With the proper redis client we can retry tasks that fail.  Since we create 9accounts at once, retry up to 10 times. Also make sure that we can lock the pwfile so that there is only one reader or writer at a time.